third_party/blink/web_tests/virtual/layout-ng-grid/external/wpt/css/css-grid/grid-definition/grid-auto-explicit-rows-001-expected.txt

This is a testharness.js-based test.
FAIL .grid 1 assert_equals: 
<div class="grid">
 <div id="c1" data-expected-width="23" data-expected-height="11"></div>
 <div id="c2" data-expected-width="52" data-expected-height="24"></div>
 <div id="c3" data-expected-width="83" data-expected-height="41"></div>
 <div id="c4" data-expected-width="120" data-expected-height="60"></div>
</div>
width expected 23 but got 784
Harness: the test ran to completion.
